+++ baybe/constraints/base.py
@@ -2,12 +2,13 @@
 
 from __future__ import annotations
 
+import gc
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any, ClassVar
 
 import pandas as pd
-from attr import define, field
-from attr.validators import ge, instance_of, min_len
+from attrs import define, field
+from attrs.validators import ge, instance_of, min_len
 
 from baybe.constraints.deprecation import structure_constraints
 from baybe.serialization import (
@@ -136,12 +137,15 @@ class ContinuousConstraint(Constraint, ABC):
 
 @define
 class CardinalityConstraint(Constraint, ABC):
-    """Abstract base class for cardinality constraints.
+    r"""Abstract base class for cardinality constraints.
 
     Places a constraint on the set of nonzero (i.e. "active") values among the
-    specified parameters, bounding it between the two given integers,
-        ``min_cardinality`` <= |{p_i : p_i != 0}| <= ``max_cardinality``
-    where ``{p_i}`` are the parameters specified for the constraint.
+    specified parameters, bounding it between the two given integers, i.e.
+
+    .. math::
+        \text{min_cardinality} \leq |\{p_i : p_i \neq 0\}| \leq \text{max_cardinality}
+
+    where :math:`\{p_i\}` are the parameters specified for the constraint.
 
     Note that this can be equivalently regarded as L0-constraint on the vector
     containing the specified parameters.
@@ -201,3 +205,6 @@ converter.register_unstructure_hook(Constraint, unstructure_base)
 # Currently affected by a deprecation
 # converter.register_structure_hook(Constraint, get_base_structure_hook(Constraint))
 converter.register_structure_hook(Constraint, structure_constraints)
+
+# Collect leftover original slotted classes processed by `attrs.define`
+gc.collect()