+++ baybe/surrogates/gaussian_process/presets/edbo.py
@@ -2,10 +2,12 @@
 
 from __future__ import annotations
 
+import gc
 from collections.abc import Collection
 from typing import TYPE_CHECKING
 
 from attrs import define
+from typing_extensions import override
 
 from baybe.kernels.basic import MaternKernel
 from baybe.kernels.composite import ScaleKernel
@@ -51,10 +53,10 @@ class EDBOKernelFactory(KernelFactory):
         * https://doi.org/10.1038/s41586-021-03213-y
     """
 
-    def __call__(  # noqa: D102
+    @override
+    def __call__(
         self, searchspace: SearchSpace, train_x: Tensor, train_y: Tensor
     ) -> Kernel:
-        # See base class.
         effective_dims = train_x.shape[-1] - len(
             [p for p in searchspace.parameters if isinstance(p, TaskParameter)]
         )
@@ -137,3 +139,7 @@ def _edbo_noise_factory(
     # OHE optimized priors
     else:
         return (GammaPrior(1.5, 0.1), 5.0)
+
+
+# Collect leftover original slotted classes processed by `attrs.define`
+gc.collect()