+++ baybe/surrogates/linear.py
@@ -2,10 +2,12 @@
 
 from __future__ import annotations
 
+import gc
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from attr import define, field
+from attrs import define, field
 from sklearn.linear_model import ARDRegression
+from typing_extensions import override
 
 from baybe.surrogates.base import IndependentGaussianSurrogate
 from baybe.surrogates.utils import batchify_mean_var_prediction, catch_constant_targets
@@ -34,12 +36,11 @@ class BayesianLinearSurrogate(IndependentGaussianSurrogate):
     _model: ARDRegression | None = field(init=False, default=None, eq=False)
     """The actual model."""
 
+    @override
     @batchify_mean_var_prediction
     def _estimate_moments(
         self, candidates_comp_scaled: Tensor, /
     ) -> tuple[Tensor, Tensor]:
-        # See base class.
-
         # FIXME[typing]: It seems there is currently no better way to inform the type
         #   checker that the attribute is available at the time of the function call
         assert self._model is not None
@@ -55,11 +56,16 @@ class BayesianLinearSurrogate(IndependentGaussianSurrogate):
 
         return mean, var
 
+    @override
     def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
-        # See base class.
         self._model = ARDRegression(**(self.model_params))
         self._model.fit(train_x, train_y.ravel())
 
+    @override
     def __str__(self) -> str:
         fields = [to_string("Model Params", self.model_params, single_line=True)]
         return to_string(super().__str__(), *fields)
+
+
+# Collect leftover original slotted classes processed by `attrs.define`
+gc.collect()