+++ baybe/surrogates/ngboost.py
@@ -4,8 +4,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from attr import define, field
+from attrs import define, field
 from ngboost import NGBRegressor
+from typing_extensions import override
 
 from baybe.parameters.base import Parameter
 from baybe.surrogates.base import IndependentGaussianSurrogate
@@ -43,28 +44,25 @@ class NGBoostSurrogate(IndependentGaussianSurrogate):
     def __attrs_post_init__(self):
         self.model_params = {**self._default_model_params, **self.model_params}
 
+    @override
     @staticmethod
     def _make_parameter_scaler_factory(
         parameter: Parameter,
     ) -> type[InputTransform] | None:
-        # See base class.
-
         # Tree-like models do not require any input scaling
         return None
 
+    @override
     @staticmethod
     def _make_target_scaler_factory() -> type[OutcomeTransform] | None:
-        # See base class.
-
         # Tree-like models do not require any output scaling
         return None
 
+    @override
     @batchify_mean_var_prediction
     def _estimate_moments(
         self, candidates_comp_scaled: Tensor, /
     ) -> tuple[Tensor, Tensor]:
-        # See base class.
-
         # FIXME[typing]: It seems there is currently no better way to inform the type
         #   checker that the attribute is available at the time of the function call
         assert self._model is not None
@@ -80,10 +78,11 @@ class NGBoostSurrogate(IndependentGaussianSurrogate):
 
         return mean, var
 
+    @override
     def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
-        # See base class.
         self._model = NGBRegressor(**(self.model_params)).fit(train_x, train_y.ravel())
 
+    @override
     def __str__(self) -> str:
         fields = [to_string("Model Params", self.model_params, single_line=True)]
         return to_string(super().__str__(), *fields)