+++ tests/test_surrogate.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 from baybe.recommenders.pure.nonpredictive.sampling import RandomRecommender
 from baybe.surrogates.gaussian_process.core import GaussianProcessSurrogate
-from baybe.utils.dataframe import add_fake_results
+from baybe.utils.dataframe import add_fake_measurements
 
 
 @patch.object(GaussianProcessSurrogate, "_fit")
@@ -12,7 +12,7 @@ def test_caching(patched, searchspace, objective):
     """A second fit call with the same context does not trigger retraining."""
     # Prepare the setting
     measurements = RandomRecommender().recommend(3, searchspace, objective)
-    add_fake_results(measurements, objective.targets)
+    add_fake_measurements(measurements, objective.targets)
     surrogate = GaussianProcessSurrogate()
 
     # First call